Move to Shared-MetadataUtilsSmi
